requirements.txt
discord.py-self
requests
async-tls-client